renamed directory and files -  old name does not reflect essence of tests.
Now all tests are inherited from base class - AddHeaderBase.

These directives and other config tests are separated.
