+++ t_modify_http_headers/__init__.py
@@ -1,0 +1,3 @@
+__all__ = ["test_common_config", "test_common_config_h2"]
+
+# vim: tabstop=8 expandtab shiftwidth=4 softtabstop=4
