--- vhost/__init__.py
@@ -1,3 +1,0 @@
-__all__ = ["test_add_hdr"]
-
-# vim: tabstop=8 expandtab shiftwidth=4 softtabstop=4
